Upgrade to dns-parser master
At the moment of writing, no new version of dns-parser is available, but
there are a lot of changes on master, some breaking, but some essential
bugfixes.

This results in one breaking change in the API: TXT records are no
longer concatenated by dns-parser, so they are also exposed separately.
To be as similar as possible to the previous API, this commit exposes
those records as a Vec<String>, rather than the String it was before.
This can be lossy if the contents is not valid UTF-8.

The new prefer_unicast argument in dns-parser was previously always
false; this commit does not change that behavior.